+++ src/hooks/contents/use-current-keyword.ts
